+++ front/package.json
@@ -1,6 +1,7 @@
 {
   "dependencies": {
     "@ant-design/icons": "^4.7.0",
+    "@types/react-router-dom": "^5.3.3",
     "antd": "^4.20.5",
     "assert": "^2.0.0",
     "axios": "^0.27.2",
